rewrite wall,ellipsoid intersections
